requirements.txt
azure-functions>=1.20.0
azure-functions-durable>=1.2.7
azure-storage-blob>=12.19.0
azure-identity>=1.15.0